+++ frontend/src/pages/App.tsx
@@ -22,14 +22,7 @@ export default function App() {
       });
       if (res.ok) {
         const user = await res.json();
-        setLoggedUser(
-          Number(user["id"]),
-          Number(user["group_id"]),
-          user["email"].toString(),
-          user["user_name"].toString(),
-          user["full_name"].toString(),
-          user["is_superuser"]
-        );
+        setLoggedUser(user);
       } else {
         navigate("/logout");
       }